# swagger.json
{
  "consumes": [
    "application/json",
    "application/xml"
  ],
  "produces": [
    "application/json",
    "application/xml"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "the purpose of this application is to provide an application\nthat is using plain go code to define an API\n\nThis should demonstrate all the possible comment annotations\nthat are available to turn go code into a fully compliant swagger 2.0 spec",
    "title": "Petstore API.",
    "termsOfService": "there are no TOS at this moment, use at your own risk we take no responsibility",
    "contact": {
      "name": "John Doe",
      "url": "http://john.doe.com",
      "email": "john.doe@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "http://opensource.org/licenses/MIT"
    },
    "version": "0.0.1"
  },
  "host": "localhost",
  "basePath": "/v2",
  "paths": {
    "/session": {
      "get": {
        "description": "Checks auth with cookie",
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "text/plain"
        ],
        "operationId": "checkAuth",
        "parameters": [
          {
            "type": "string",
            "description": "Name to be returned.",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The hello message"
          }
        }
      },
      "post": {
        "description": "Logins",
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "text/plain"
        ],
        "operationId": "signIn",
        "parameters": [
          {
            "type": "string",
            "description": "Name to be returned.",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The hello message"
          }
        }
      },
      "delete": {
        "description": "LogOut and delete cookie",
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "text/plain"
        ],
        "operationId": "LogOut",
        "parameters": [
          {
            "type": "string",
            "description": "Name to be returned.",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The hello message"
          }
        }
      }
    },
    "/user": {
      "get": {
        "description": "Return list of leaders",
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "text/plain"
        ],
        "operationId": "LeaderBoard",
        "parameters": [
          {
            "type": "int",
            "description": "Number of page to show",
            "name": "page",
            "in": "query",
            "required": true
          },
          {
            "type": "int",
            "description": "Size of single page",
            "name": "page_size",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The hello message"
          }
        }
      },
      "put": {
        "description": "Updates user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "UpdateUser",
        "parameters": [
          {
            "type": "int",
            "description": "Number of page to show",
            "name": "page",
            "in": "query",
            "required": true
          },
          {
            "type": "int",
            "description": "Size of single page",
            "name": "page_size",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The hello message"
          }
        }
      },
      "post": {
        "description": "Register new user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "RegisterUser",
        "parameters": [
          {
            "type": "int",
            "description": "Number of page to show",
            "name": "page",
            "in": "query",
            "required": true
          },
          {
            "type": "int",
            "description": "Size of single page",
            "name": "page_size",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The hello message"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "KEY",
      "in": "header"
    },
    "oauth2": {
      "type": "oauth2",
      "in": "header",
      "flow": "accessCode",
      "authorizationUrl": "/oauth2/auth",
      "tokenUrl": "/oauth2/token",
      "scopes": {
        "bar": "foo"
      }
    }
  },
  "security": [
    {
      "api_key": []
    }
  ],
  "x-meta-array": [
    "value1",
    "value2"
  ],
  "x-meta-array-obj": [
    {
      "name": "obj",
      "value": "field"
    }
  ],
  "x-meta-value": "value"
}